Module err_context::prelude

A module with reexports to wildcard-import all relevant traits.

The library adds methods to existing types by extension traits. For them to work, they need to be in scope. It is more convenient to import them all from prelude instead of individually.

use err_context::prelude::*;

Only the traits are imported and they are imported anonymously (so their names can't clash with anything, but they also can't be referred directly).

Re-exports

pub use crate::BoxedErrorExt as _;

pub use crate::BoxedResultExt as _;

pub use crate::ErrorExt as _;

pub use crate::ResultExt as _;